+++ ICD.Profound.ConnectPROCommon/SettingsTree/ConnectProRootSettingsNode.cs
@@ -2,7 +2,6 @@
 using ICD.Profound.ConnectPROCommon.Rooms;
 using ICD.Profound.ConnectPROCommon.SettingsTree.About;
 using ICD.Profound.ConnectPROCommon.SettingsTree.Administrative;
-using ICD.Profound.ConnectPROCommon.SettingsTree.Conferencing;
 using ICD.Profound.ConnectPROCommon.SettingsTree.CUE;
 using ICD.Profound.ConnectPROCommon.SettingsTree.RoomCombine;
 using ICD.Profound.ConnectPROCommon.SettingsTree.TouchFreeConferencing;
@@ -24,7 +23,6 @@ namespace ICD.Profound.ConnectPROCommon.SettingsTree
 		{
 			yield return new AboutSettingsNode();
 			yield return new AdministrativeSettingsNode();
-			yield return new ConferencingSettingsNode();
 			yield return new CueSettingsNode();
 			yield return new RoomCombineSettingsNode();
 			yield return new TouchFreeSettingsNode();