+++ CHANGELOG.md
@@ -17,6 +17,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Changed
  - Meetings will automatically timeout if the room is unoccupied for 10 minutes.
 
+### Removed
+ - Removed "directory" settings menu item and subpages.
+
 ## [14.1.2] - 2020-09-28
 ### Changed
  - Fixed room "Not In Meeting" activity priority